Integrity check - GAME COMPLETE
Bug fixing time T-T

+++ Engine/Header Files/Components/Weight.h
@@ -1,4 +1,4 @@
 struct Weight
 {
-	float weight;
+	int weight;
 };
